move native hints to packages they are "used"
