+++ src/main/java/tech/jhipster/lite/module/infrastructure/primary/NativeHints.java
@@ -1,9 +1,13 @@
 package tech.jhipster.lite.module.infrastructure.primary;
 
-import org.springframework.aot.hint.*;
-import tech.jhipster.lite.module.domain.landscape.*;
+import org.springframework.aot.hint.MemberCategory;
+import org.springframework.aot.hint.RuntimeHints;
+import org.springframework.aot.hint.RuntimeHintsRegistrar;
+import tech.jhipster.lite.common.domain.Generated;
+import tech.jhipster.lite.module.domain.landscape.JHipsterLandscapeElementType;
 
-public class NativeHints implements RuntimeHintsRegistrar {
+@Generated(reason = "Not testing native runtime hints")
+class NativeHints implements RuntimeHintsRegistrar {
 
   @Override
   public void registerHints(RuntimeHints hints, ClassLoader classLoader) {
