+++ src/main/java/tech/jhipster/lite/module/infrastructure/secondary/git/NativeJGitConfiguration.java
@@ -1,0 +1,10 @@
+package tech.jhipster.lite.module.infrastructure.secondary.git;
+
+import org.springframework.context.annotation.*;
+import tech.jhipster.lite.common.domain.*;
+
+@ImportRuntimeHints(NativeHints.class)
+@Generated(reason = "Not testing native configuration")
+@Configuration
+class NativeJGitConfiguration {
+}
