+++ src/main/java/tech/jhipster/lite/module/infrastructure/secondary/javadependency/NativeJavaDependencyConfiguration.java
@@ -1,0 +1,11 @@
+package tech.jhipster.lite.module.infrastructure.secondary.javadependency;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.ImportRuntimeHints;
+import tech.jhipster.lite.common.domain.Generated;
+
+@ImportRuntimeHints(NativeHints.class)
+@Generated(reason = "Not testing native configuration")
+@Configuration
+public class NativeJavaDependencyConfiguration {
+}
