+++ src/main/java/tech/jhipster/lite/project/infrastructure/secondary/ProjectFormatterConfiguration.java
@@ -2,7 +2,11 @@ package tech.jhipster.lite.project.infrastructure.secondary;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.context.annotation.*;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.ImportRuntimeHints;
+import org.springframework.context.annotation.Bean;
+
 import tech.jhipster.lite.common.domain.Generated;
 
 @Configuration
